api-gateway/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
httpx
python-multipart
